+++ packages/core-js-sdk/src/sdk/types.ts
@@ -50,9 +50,10 @@ export type UserResponse = User & {
   roleNames?: string[];
   userTenants?: UserTenant[];
   createTime: number;
-  totp: boolean;
-  saml: boolean;
-  oauth?: Record<string, boolean>;
+  TOTP: boolean;
+  SAML: boolean;
+  OAuth?: Record<string, boolean>;
+  customAttributes?: Record<string, any>;
 };
 
 /** A tenant association mapping  */
@@ -203,6 +204,12 @@ export type FlowResponse = {
   redirect?: {
     url: string;
   };
+  // SAML IDP response (this will be used to build the html form response goes from the IDP through the end user browser to the SP)
+  samlIdpResponse?: {
+    url: string;
+    samlResponse: string;
+    relayState: string;
+  };
   // webauthn data - if action is one of 'webauthnCreate', 'webauthnGet'
   webauthn?: {
     transactionId: string;
@@ -222,11 +229,16 @@ export type FlowResponse = {
 
 export type Options = {
   redirectUrl?: string;
+  location?: string;
   tenant?: string;
   deviceInfo?: DeviceInfo;
   lastAuth?: LastAuth;
   redirectAuth?: RedirectAuth;
   oidcIdpStateId?: string;
+  preview?: boolean;
+  samlIdpStateId?: string;
+  samlIdpUsername?: string;
+  ssoAppId?: string;
 };
 
 export type ResponseData = Record<string, any>;